exchanged the List for a HashMap in TransactionHistory

+++ src/main/java/infrastructure/TransactionService.java
@@ -1,6 +1,6 @@
 package infrastructure;
 import domen.Player;
 public interface TransactionService {
-    boolean debit(Player player, long amount, String transactionId);
-    boolean credit(Player player, long amount, String transactionId);
+    boolean debit(Player player, int amount, String transactionId);
+    boolean credit(Player player, int amount, String transactionId);
 }
